feat(refactor): modularize booking card dialog actions on My booking tab
