+++ apps/web/src/components/guest/my-bookings/booking-card/booking-card-header.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+import { PropertyInfo } from "@/components/guest/my-bookings/property-info";
+import { useBookingCardContext } from "./booking-card-context";
+
+export const BookingCardHeader = () => {
+  const { booking } = useBookingCardContext();
+
+  return <PropertyInfo property={booking.Property} room={booking.Room} />;
+};
