+++ apps/web/src/components/guest/my-bookings/booking-card/index.tsx
@@ -1,0 +1,3 @@
+// apps/web/src/components/guest/my-bookings/booking-card/index.tsx
+export { BookingCard } from "./booking-card";
+export { useBookingCardContext } from "./booking-card-context";
